handle swipe to delete by showing confirmation bottom sheet before deleting task then delete
